Add classes and fix main
